requirements.txt
flask
flask-cors
requests
pytest
google-genai
python-dotenv
